+++ nodejs/awsx/ecs/cluster.ts
@@ -145,6 +145,8 @@ export class Cluster
     }
 }
 
+(<any>Cluster.prototype.createAutoScalingGroup).doNotCapture = true;
+
 // The shape we want for ClusterArgs.  We don't export this as 'Overwrite' types are not pleasant to
 // work with. However, they internally allow us to succinctly express the shape we're trying to
 // provide. Code later on will ensure these types are compatible.